features: add a script to restore the workspace to a given state
All the state-saving is really there to ease the development of the
features. The obvious target is to avoid re-bootstrapping, but even
without this, running all features takes quite a bit of time (and
they are sequential).

The idea is to store the state of select directories as git tags and
restore the state as it was. The tags represent the end state of a
scenario. The features/restore_aruba_state script is given a feature
and scenario and will restore to the state to what it was before. I.e.
one runs restore_aruba_state FEATURE SCENARIO before running cucumber
on the feature and scenario.

The feature is given as a file. Without a scenario argument, the script
will take the first scenario of the file.